Refactoring of new bepias sub-views to use them for both instruction and visa
