+++ frontend/src/components/NewBepiasViews/BepiasSidebar.vue
