+++ frontend/src/components/NewBepiasViews/HistorySection/index.vue
