+++ frontend/src/components/NewBepiasViews/ProductSection/InstructionResults.vue
