+++ frontend/src/components/NewBepiasViews/ProductSection/LastComment.vue
