+++ frontend/src/components/NewBepiasViews/ProductSection/index.vue
