+++ frontend/src/components/NewBepiasViews/SectionHeader.vue
